fix(core): :bug: removed unwanted code and made the code structured
removed unwanted code and made the code structured for the users to see and use

ref: #97
